# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Peer Rater Api Documentation",
        "description": "Peer Rater Api Documentation",
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/categories/{category}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "get a category infomation by ID",
                "description": "Returns an array contains all the categories infomation",
                "operationId": "categories.show",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "category ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 12
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Invalid input"
                    }
                }
            },
            "put": {
                "tags": [
                    "Category"
                ],
                "summary": "update a category description",
                "description": "Returns the updated categories infomation",
                "operationId": "categories.update",
                "parameters": [
                    {
                        "name": "description",
                        "in": "path",
                        "description": "category description",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "This is a description"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Invalid input"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "delete a category by Category ID",
                "description": "Return the Category infomation",
                "operationId": "categories.destroy",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "delete category by category ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
        "/api/peer-groups/{peerGroup}": {
            "get": {
                "tags": [
                    "PeerGroup"
                ],
                "summary": "get the peerGroup infomation",
                "description": "Returns the peer group infomation",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "peerGroup",
                        "in": "path",
                        "description": "get peerGroup information by ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Invalid input"
                    }
                }
            },
            "put": {
                "tags": [
                    "PeerGroup"
                ],
                "summary": "update the peerGroup infomation",
                "description": "Returns the peer group infomation",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "peerGroup",
                        "in": "path",
                        "description": "get peerGroup information by ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "description",
                        "in": "path",
                        "description": "get peerGroup information by ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Invalid input"
                    }
                }
            },
            "delete": {
                "tags": [
                    "PeerGroup"
                ],
                "summary": "delete the peerGroup by peerGroup ID",
                "description": "Returns the peer group infomation",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "peerGroup",
                        "in": "path",
                        "description": "get peerGroup information by ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
        "/api/peer-groups/{peerGroup_id}/surveys/{survey_id}/ratings/{rating_id}": {
            "get": {
                "tags": [
                    "Rating"
                ],
                "summary": "get a rating infomation",
                "description": "(Not Confirm) Returns the rating infomation by rating id, survey id and peerGroup id",
                "operationId": "ratings.show",
                "parameters": [
                    {
                        "name": "survey_id",
                        "in": "path",
                        "description": "survey ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "peerGroup_id",
                        "in": "path",
                        "description": "peerGroup ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "rating_id",
                        "in": "path",
                        "description": "This is the rating ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success operation"
                    },
                    "500": {
                        "description": "Invalid input"
                    }
                }
            },
            "put": {
                "tags": [
                    "Rating"
                ],
                "summary": "update a rating",
                "description": "Update the rating infomation by rating id, survey id and peerGroup id",
                "operationId": "ratings.update",
                "parameters": [
                    {
                        "name": "survey_id",
                        "in": "path",
                        "description": "survey ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "peerGroup_id",
                        "in": "path",
                        "description": "peerGroup ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "rating_id",
                        "in": "path",
                        "description": "rating ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "rating",
                        "in": "path",
                        "description": "rating content",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success operation"
                    },
                    "500": {
                        "description": "Invalid input"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Rating"
                ],
                "summary": "delete a rating",
                "description": "delete the rating by rating id, survey id and peerGroup id",
                "operationId": "ratings.destroy",
                "parameters": [
                    {
                        "name": "survey_id",
                        "in": "path",
                        "description": "survey ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "peerGroup_id",
                        "in": "path",
                        "description": "peerGroup ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "rating_id",
                        "in": "path",
                        "description": "rating ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success operation"
                    },
                    "500": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/peer-groups/{peerGroup_id}/surveys/": {
            "get": {
                "tags": [
                    "Survey"
                ],
                "summary": "get the Survey group by peerGroup id",
                "description": "Returns the Survey infomation as an array",
                "operationId": "Survey.index",
                "parameters": [
                    {
                        "name": "peerGroup_id",
                        "in": "path",
                        "description": "peerGroup ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User not exsits! / Success, User is detached from peer group / Error, User does not exsit in peer group"
                    },
                    "500": {
                        "description": "Invalid input"
                    }
                }
            },
            "post": {
                "tags": [
                    "Survey"
                ],
                "summary": "create a Survey",
                "description": "create a new Survey by accepting peerGroup_id",
                "operationId": "Survey.store",
                "parameters": [
                    {
                        "name": "peerGroup_id",
                        "in": "path",
                        "description": "peerGroup ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success operation"
                    },
                    "500": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/peer-groups/{peerGroup_id}/surveys/{survey_id}": {
            "get": {
                "tags": [
                    "Survey"
                ],
                "summary": "show a Survey infomation",
                "description": "show a new Survey by accepting peerGroup_id and survey_id",
                "operationId": "Survey.show",
                "parameters": [
                    {
                        "name": "peerGroup_id",
                        "in": "path",
                        "description": "peerGroup ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "survey_id",
                        "in": "path",
                        "description": "survey ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success operation"
                    },
                    "500": {
                        "description": "Invalid input"
                    }
                }
            },
            "put": {
                "tags": [
                    "Survey"
                ],
                "summary": "update a Survey infomation",
                "description": "update a new Survey by accepting peerGroup_id, survey_id and a new peer_group_id",
                "operationId": "Survey.update",
                "parameters": [
                    {
                        "name": "peerGroup_id",
                        "in": "path",
                        "description": "current peerGroup ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "survey_id",
                        "in": "path",
                        "description": "survey ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "peer_group_id",
                        "in": "path",
                        "description": "updated peerGroup ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success operation"
                    },
                    "500": {
                        "description": "Invalid input"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Survey"
                ],
                "summary": "delete a Survey infomation",
                "description": "delete a new Survey by accepting peerGroup_id, survey_id",
                "operationId": "Survey.destroy",
                "parameters": [
                    {
                        "name": "peerGroup_id",
                        "in": "path",
                        "description": "current peerGroup ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "survey_id",
                        "in": "path",
                        "description": "survey ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success operation"
                    },
                    "500": {
                        "description": "Invalid input"
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    },
...
}